bin/requirements.txt
setuptools
solaredge
paho-mqtt